+++ src/components/projectDetail/supplier/purchaseOrder/supplierPoReport.vue
@@ -1,0 +1,306 @@
+<template>
+<div id="supplierPoLetterForm">
+
+    <div class="container">
+      <div class="card" style="margin-top: 50px; margin-bottom: 100px;">
+        <h3 class="card-header">
+          Supplier Purchase Order Detail
+          <router-link to="/spo/list" class="btn btn-danger ml-3" style="float: right;">Back</router-link>
+          <button
+            id="printInvoice"
+            type="button"
+            class="btn btn-primary ml-3"
+            style="float: right;"
+            @click="printInvoice"
+          >
+            <i class="fa fa-print"></i> Print
+          </button>
+          <button type="button" class="btn btn-primary ml-3" style="float: right;">
+            <i class="fa fa-file-pdf-o"></i> Export
+          </button>
+        </h3>
+
+        <div class="card-body">
+        
+        <div class="header mb-5">
+            <div class="container">
+                <div class="row">
+                    <div class="col"></div>
+                    <div class="col-5 text-center">
+                        <div class="row">
+                            <h6 class="text-danger text-uppercase">GRIZPOWER SDN BHD (713693-T)</h6>
+                            <div>No 18 & 18-1, Jalan Seti Rejang 3</div>
+                            <div>Taman Sri Rampai</div>
+                            <div>53300 Kuala Lumpur, Malaysia</div>
+                        </div>
+                    </div>
+                    <div class="col text-right">
+                        <div class="row align-items-end">
+                            <p>
+                                Tel: 603-4144 3365<br/>
+                                Fax: 603-4144 3430<br/>
+                                E-Mail: info@grizpower.com
+                            </p>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+        <div class="contact">
+            <div class="body">
+
+                    <div class="row mb-3">
+                      <div class="col">
+                        <h5>XXXXX XXXXXXX XXXX Sdn Bhd</h5>
+                        <p>
+                            18th KM<br/>
+                            P.O.<br/>
+                            97008 Bintulu, Sarawak, Malaysia
+                        </p>
+                      </div>
+                      <div class="col">
+                        <div class="row">
+                            <div class="col">
+                                <b>Date of Quote :</b>
+                            </div>
+                            <div class="col">
+                                01st Jan 2019
+                            </div>
+                        </div>
+                      </div>
+                    </div>
+
+                    <div class="row mb-3">
+                      <div class="col">
+                        <div class="row">
+                            <div class="col-4">
+                                <b>Attention :</b>
+                            </div>
+                            <div class="col">
+                                XXXXXXXXXXXXX
+                            </div>
+                        </div>
+                        <div class="row">
+                            <div class="col-4">
+                                <b>Contact :</b>
+                            </div>
+                            <div class="col">
+                                086-XXX XXXX
+                            </div>
+                        </div>
+                        <div class="row">
+                            <div class="col-4">
+                                <b>E-Mail :</b>
+                            </div>
+                            <div class="col">
+                                scholastica.siman@petronas.com.my
+                            </div>
+                        </div>
+                      </div>
+                      <div class="col">
+                        <div class="row">
+                            <div class="col-4">
+                                <b>Prepared by :</b>
+                            </div>
+                            <div class="col">
+                                XXXXXXXXXXXXX
+                            </div>
+                        </div>
+                        <div class="row">
+                            <div class="col-4">
+                                <b>Contact :</b>
+                            </div>
+                            <div class="col">
+                                086-XXX XXXX
+                            </div>
+                        </div>
+                        <div class="row">
+                            <div class="col-4">
+                                <b>E-Mail :</b>
+                            </div>
+                            <div class="col">
+                                sales@grizpower.com.my
+                            </div>
+                        </div>
+                      </div>
+                    </div>
+
+
+                    <div class="row mb-3">
+                        <div class="col">
+                          <div class="row">
+                              <div class="col-5">
+                                  <b>RFQ No. :</b>
+                              </div>
+                              <div class="col">
+                                  XXXXXXXXXXXXX
+                              </div>
+                          </div>
+                          <div class="row">
+                              <div class="col-5">
+                                  <b>RFQ Name :</b>
+                              </div>
+                              <div class="col">
+                                  086-XXX XXXX
+                              </div>
+                          </div>
+                          <div class="row">
+                              <div class="col-5">
+                                  <b>SUS Portal ID :</b>
+                              </div>
+                              <div class="col">
+                                  GRIZPOWER
+                              </div>
+                          </div>
+                        </div>
+                        <div class="col">
+                          <div class="row">
+                              <div class="col-5">
+                                  <b>GP Ref No :</b>
+                              </div>
+                              <div class="col">
+                                  GP3582
+                              </div>
+                          </div>
+                          <div class="row">
+                              <div class="col-5">
+                                  <b>Petronas License :</b>
+                              </div>
+                              <div class="col">
+                                  713693-T
+                              </div>
+                          </div>
+                          <div class="row">
+                              <div class="col-5">
+                                  <b>Closing Date :</b>
+                              </div>
+                              <div class="col">
+                                  18th May 2019
+                              </div>
+                          </div>
+                        </div>
+                      </div>
+
+                      <div class="row">
+                        <div class="col">
+                            <h6 class="text-danger">GP3582 - TECHNICAL PROPOSAL (UNPRICED)</h6>
+                        </div>
+                        <div class="col">
+                          <div class="row">
+                              <div class="col">
+                                  <b>CURRENCY :</b>
+                              </div>
+                              <div class="col">
+                                  MALAYSIAN RINGGIT (MYR)
+                              </div>
+                          </div>
+                        </div>
+                      </div>
+
+            </div>
+        </div>
+
+        <hr>
+
+        <div class="items">
+            <table class="table table-bordered">
+                <thead>
+                  <tr>
+                    <th scope="col">#</th>
+                    <th scope="col">DESCRIPTION</th>
+                    <th scope="col">QTY</th>
+                    <th scope="col">UNIT PRICE</th>
+                    <th scope="col">TOTAL PRICE</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  <tr>
+                    <th scope="row">1</th>
+                    <td>Item Description</td>
+                    <td>12</td>
+                    <td>2,500.00</td>
+                    <td>9,999.00</td>
+                  </tr>
+                  <tr>
+                    <th scope="row">2</th>
+                    <td>Item Description</td>
+                    <td>4</td>
+                    <td>8,000.00</td>
+                    <td>9,999.00</td>
+                  </tr>
+                  <tr>
+                    <th scope="row">3</th>
+                    <td>Item Description</td>
+                    <td>5</td>
+                    <td>2,000.00</td>
+                    <td>9,999.00</td>
+                  </tr>
+                  <tr>
+                    <th scope="row"></th>
+                    <td colspan="3"><h6>GRAND TOTAL IN MYR</h6></td>
+                    <td><h6>29,900.00</h6></td>
+                  </tr>
+                </tbody>
+            </table>
+        </div>
+
+        <hr>
+
+        <div class="info mb-5 mt-5">
+            <div class="row mb-3">
+                <div class="col"><u><b>TERMS AND CONDITIONS</b></u></div>
+            </div>
+            <div class="row mb-3">
+                <div class="col"><b>1. PRICE VALIDITY: </b></div>
+                <div class="col-8">30 Days from date of submission / closing date as stated above. </div>
+            </div>
+            <div class="row mb-3">
+                <div class="col"><b>2. PAYMENT TERMS: </b></div>
+                <div class="col-8">30 days after acceptance of delivery and receipt of invoice. </div>
+            </div>
+            <div class="row mb-3">
+                <div class="col"><b>3. DELIVERY TERMS: </b></div>
+                <div class="col-8">Delivery At Place (DAP),  <b>ABF Warehouse, 97008 Bintulu, Malaysia</b> </div>
+            </div>
+            <div class="row mb-3">
+                <div class="col"><b>4. DELIVERY CHARGES: </b></div>
+                <div class="col-8">Includes factory packing charges, handling fees, ground and air freight charges + Insurance & any custom duty payable in or from the country of origin and Malaysia with final delivery directly to the address stated above. </div>
+            </div>
+            <div class="row mb-3">
+                <div class="col"><b>5. LEAD TIME: </b></div>
+                <div class="col-8"><b>42 days (6 weeks)</b> after receipt of order via SUS Portal. The lead time is strictly from the date of order to the date these goods are ready for collection at the address stated above. Any delays for the collection and/or arrangements of these items by your appointed freight forwarder should not reflect or impact the delivery performance or Purchase Order obligations of Grizpower. </div>
+            </div>
+        </div>
+
+        <div class="sign">
+            <div class="row mb-5">
+                <div class="col">For and on behalf of Grizpower Sdn. Bhd.</div>
+            </div>
+            <div class="row mt-5">
+                <div class="col">...........................................................................</div>
+            </div>
+            <div class="row">
+                <div class="col">
+                    XXX XXXX<br/>
+                    SENIOR SALES EXECUTIVE<br/>
+                    +6019-XXX XXXX
+                </div>
+            </div>
+        </div>
+
+        </div>
+    </div>
+    </div>
+</div>
+</template>
+
+<script>
+export default {
+    name: 'supplierPoLetterForm'
+}
+</script>
+
+<style scoped>
+
+</style>
